+++ CHANGE_LOG.md
@@ -13,6 +13,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 
 - Fixed an issue where opening a chat right after agent ends the previous chat doesn't work
+- Fixed an issue where opening/refreshing a persistent chat shows the Reconnect Pane
 
 ## [1.2.1] - 2023-7-24